Fix error validation when add Profit to City

+++ app/Http/Requests/Sameleon/City/CityFormRequest.php
@@ -28,7 +28,7 @@ class CityFormRequest extends FormRequest
             'has_profit' => ['nullable', Rule::in([1, '1', true, 'on', 'yes', 'oui', '0', 'no', 'non', false])],
             'name' => ['required', 'string', 'unique:cities'],
             'frais' => ['required', 'numeric'],
-            'profit' => ['nullable', 'required_if:has_profit,true', 'float'],
+            'profit' => ['nullable', 'required_if:has_profit,true', 'numeric'],
         ];
     }
 }
